CRUD CICLOS LISTOS

+++ src/main/java/pe/com/edu/prismaapp/prisma/repositories/CycleRepository.java
@@ -3,5 +3,5 @@ package pe.com.edu.prismaapp.prisma.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.com.edu.prismaapp.prisma.entities.Cycle;
 
-public interface CycleRepository extends JpaRepository<Cycle, Integer> {
+public interface CycleRepository extends JpaRepository<Cycle, Long> {
 }
